Move dependency file to dataflow runner directory
