+++ sdks/python/apache_beam/utils/profiler.py
@@ -24,10 +24,8 @@ import pstats
 import StringIO
 import tempfile
 import time
-from threading import Timer
 import warnings
-
-from apache_beam.utils.dependency import _dependency_file_copy
+from threading import Timer
 
 
 class Profile(object):
@@ -35,11 +33,13 @@ class Profile(object):
 
   SORTBY = 'cumulative'
 
-  def __init__(self, profile_id, profile_location=None, log_results=False):
+  def __init__(self, profile_id, profile_location=None, log_results=False,
+               file_copy_fn=None):
     self.stats = None
     self.profile_id = str(profile_id)
     self.profile_location = profile_location
     self.log_results = log_results
+    self.file_copy_fn = file_copy_fn
 
   def __enter__(self):
     logging.info('Start profiling: %s', self.profile_id)
@@ -51,14 +51,14 @@ class Profile(object):
     self.profile.disable()
     logging.info('Stop profiling: %s', self.profile_id)
 
-    if self.profile_location:
+    if self.profile_location and self.file_copy_fn:
       dump_location = os.path.join(
           self.profile_location, 'profile',
           ('%s-%s' % (time.strftime('%Y-%m-%d_%H_%M_%S'), self.profile_id)))
       fd, filename = tempfile.mkstemp()
       self.profile.dump_stats(filename)
       logging.info('Copying profiler data to: [%s]', dump_location)
-      _dependency_file_copy(filename, dump_location)  # pylint: disable=protected-access
+      self.file_copy_fn(filename, dump_location)  # pylint: disable=protected-access
       os.close(fd)
       os.remove(filename)
